Upgrade dependency size to 0.4
`size` 0.4 uses the builder api for manually formatting sizes so that only the
parameters you wish to change (in this case, `Base` but not `Style`) need to be
listed.

This patch should only be merged after gorilla-devs/libium#8 is merged and this
crate is updated to use that release of libium, to keep the `size` crates
aligned.